requirements.txt
# Core AI/ML
openai>=1.12.0
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5

# Market Data & Analysis
yfinance>=0.2.36
TA-Lib>=0.4.28
finnhub-python>=2.4.19
pandas>=2.0.0
numpy>=1.24.0

# News & Web
newsapi-python>=0.2.7
requests>=2.31.0
beautifulsoup4>=4.12.0

# Document Processing
pypdf2>=3.0.0
openpyxl>=3.1.0

# Vector Storage & Memory
chromadb>=0.4.22
sentence-transformers>=2.3.1

# Risk Models
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.2
pydantic>=2.5.0